LayoutTests/imported/w3c/web-platform-tests/html/webappapis/scripting/processing-model-2/window-onerror-with-cross-frame-event-listeners-1-expected.txt

CONSOLE MESSAGE: ReferenceError: Can't find variable: thereIsNoSuchCallable


FAIL The error event from an event listener should fire on that listener's global assert_true: expected true got false
